requirements.txt
alembic==1.8.1
Babel==2.11.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
WTForms==3.0.1
